requirements.txt
azure-core==1.30.2
azure-servicebus==7.12.2
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
isodate==0.6.1
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
